Move matching to utils
