+++ src/modules/Driver.ts
@@ -1,4 +1,4 @@
-import { ElementDataObject, XMLAttributes, AppUIResponseObject, Params, Action } from '../types/RokulRunnings';
+import { ElementDataObject, AppUIResponseObject, Params, Action } from '../types/RokulRunnings';
 import * as http from '../utils/http';
 import { sleep } from '../utils/sleep';
 import { Buttons } from './RokulRunnings';
@@ -7,6 +7,7 @@ import * as indigestion from 'indigestion';
 import { generateHeaders } from '../utils/authHeaders';
 import { jsonFormatterDeviceInfo, jsonFormatterMediaPlayer, jsonFormatterApps } from '../utils/formatters';
 import { xmlParser as parser } from '../utils/xmlParser';
+import { matchElements } from '../utils/elementMatching';
 
 export class Driver {
   constructor(public rokuIPAddress: string, public retries: number) {
@@ -174,57 +175,6 @@ export class Driver {
     return responseArray;
   }
 
-  /** Finds all elements that match the specified search terms */
-  matchElements(searchElements: ElementDataObject, actualElement: AppUIResponseObject, elementName: string) {
-    const foundElements = [];
-    // Iterate over each key in the passed in element
-    for (const key of Object.keys(actualElement)) {
-      if (typeof actualElement[key] === 'object') {
-        if (Array.isArray(actualElement[key])) {
-          const elementArray = actualElement[key];
-          for (let i = 0; i < elementArray.length; i++) {
-            // If `actualElement[key]` is an Array object, each element needs to be iterated over
-            foundElements.push(...this.matchElements(searchElements, actualElement[key][i], key));
-          }
-        } else {
-          if (key === 'attributes' && this.checkElements({ searchElements, actualElement: actualElement[key] })) {
-            // This covers the cases where `searchElements.using` === 'text' or 'attribute'
-            foundElements.push({ [elementName]: actualElement });
-          } else if (key !== 'attributes' && this.checkElements({ searchElements, elementTag: key })) {
-            // This covers the case where `searchElements.using` === 'tag'
-            foundElements.push({ [key]: actualElement[key] });
-          } else {
-            // If the above conditions are not met, then the element needs to be passed through this function
-            foundElements.push(...this.matchElements(searchElements, actualElement[key] as AppUIResponseObject, key));
-          }
-        }
-      }
-    }
-
-    return foundElements;
-  }
-
-  /** Determines if a singular element matches the search criteria */
-  checkElements({
-    searchElements,
-    actualElement = {},
-    elementTag = '',
-  }: {
-    searchElements: ElementDataObject;
-    actualElement?: XMLAttributes;
-    elementTag?: string;
-  }) {
-    if (searchElements.using === 'text' && actualElement.text) {
-      return searchElements.value === actualElement.text;
-    } else if (searchElements.using === 'attr' && actualElement[searchElements.attribute]) {
-      return searchElements.value === actualElement[searchElements.attribute];
-    } else if (searchElements.using === 'tag') {
-      return searchElements.value === elementTag;
-    } else {
-      return false;
-    }
-  }
-
   /** Searches for an element on the page, starting from the screen root. The first located element will be returned as a WebElement JSON object. */
   async getUIElement({ data, retries }: { data: ElementDataObject; retries: number }) {
     const elements = await this.getUIElements({ data, retries });
@@ -251,7 +201,7 @@ export class Driver {
     if (!sceneName || !elements) {
       throw new Error('Cannot find root Scene component');
     }
-    const foundElements = this.matchElements(data, elements, sceneName);
+    const foundElements = matchElements(data, elements, sceneName);
 
     return foundElements;
   }
